Fix release-build Dam hang (path table UB) and curved pointer (game atan2f)
- init_path_table_links wrote its cursor through validationGroupCursors[-3];
  the optimised build looped forever loading the Dam. Plain local now; same
  for gunfire.c's KF7 flash scale in stackpad2[-8]
- The game's own atan2f (0..2pi), acosf and asinf override libm for the VR
  code; the curved screen's left half mapped to u ~7. VR code uses the double
  versions
- Pointer miss probe and START/picker test hooks for the launcher

+++ port/vr/vr_input.cpp
@@ -972,7 +972,7 @@ void QuatSlerp(const float a[4], const float b[4], float t, float out[4])
         return;
     }
 
-    float theta0 = acosf(dot);
+    float theta0 = (float)acos((double)dot);   // not the game's acosf (math_asinfacosf.c)
     float theta  = theta0 * t;
     float sinT0  = sinf(theta0);
     float sinT   = sinf(theta);
